+++ R/internal_functions.R
@@ -45,3 +45,17 @@
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(.getPackageVersion(pkgname))
 }
+
+# next function allows us to update the object
+.byRef <- function(..., envir=parent.frame(), inherits=TRUE) {
+  cl <- match.call(expand.dots = TRUE)
+  cl[c(1, match(c("envir", "inherits"), names(cl), 0L))] <- NULL
+  for (x in as.list(cl)) {
+    s <- substitute(x)
+    sx <- do.call(substitute, list(s), envir=envir)
+    dx <- deparse(sx)
+    expr <- substitute(assign(dx, s, envir=parent.frame(), inherits=inherits))
+    do.call(on.exit, list(expr, add=TRUE), envir=envir)
+  }
+}
+